owner service impl completed

+++ src/main/java/com/klip/petcare/repository/jpa/core/OwnerRepository.java
@@ -3,5 +3,5 @@ package com.klip.petcare.repository.jpa.core;
 import com.klip.petcare.entity.core.OwnerEntity;
 import com.klip.petcare.repository.jpa.base.GenericRepository;
 
-public interface OwnerRepository extends GenericRepository<OwnerEntity, Long> {
+public interface OwnerRepository extends GenericRepository<OwnerEntity, Integer> {
 }
